fashionfolio-backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]
sqlalchemy>=2.0.0
psycopg2-binary
pydantic>=2.0.0
httpx
python-dotenv
python-jose[cryptography]